requirements.txt
setuptools
finvizfinance==1.1.0
numpy==1.26.4
pandas==2.2.3
pandas_ta==0.3.14b0
plotly==5.24.1
pygooglenews==0.1.3
pygwalker==0.4.9.13
pymongo==4.11
python_dateutil==2.9.0.post0
streamlit==1.41.1
yfinance==0.2.50
